Prevent rebuild of API extension module failing when target missing.
Build rules will now only attempt to remove API .so file if present,
preventing the setup script from failing when the library is missing.

+++ control/setup.py
@@ -64,7 +64,8 @@ class ExcaliburBuildExt(build_ext):
                 log.info("API library {} was recompiled, forcing rebuild of extension".format(
                     fem_api.name
                 ))
-                os.remove(self.get_ext_fullpath(fem_api.name))
+                if os.path.exists(self.get_ext_fullpath(fem_api.name)):
+                    os.remove(self.get_ext_fullpath(fem_api.name))
            
         # Run the real build_ext command
         build_ext.run(self)
